+++ src/dev-tools.test.ts
@@ -1,7 +1,9 @@
+import {assert} from "chai";
 import {globalStateChanges$} from "./dev-tools";
 import {enableTyduxDevelopmentMode} from "./development";
 import {Mutators} from "./mutators";
 import {Store} from "./Store";
+import {createAsyncPromise} from "./test-utils";
 
 describe("DevTools", function () {
 
@@ -36,7 +38,7 @@ describe("DevTools", function () {
         }
 
         globalStateChanges$.subscribe((event) => {
-        	eventActionTypes.push(event.storeName + " # " + event.action.type);
+            eventActionTypes.push(event.storeName + " # " + event.action.type);
         });
 
         const store = new MyStore("myStore", new MyMutators(), {});
@@ -63,5 +65,59 @@ describe("DevTools", function () {
         }, 0);
 
     });
+
+    it("events contain the store and mutators name, support for async methods", function (done) {
+        const eventActionTypes: string[] = [];
+
+        class MyMutators extends Mutators<any> {
+            mut1() {
+            }
+
+            mut2() {
+            }
+        }
+
+        class MyStore extends Store<MyMutators, any> {
+
+            async action1() {
+                this.mutate.mut1();
+                this.innerAction();
+                this.mutate.mut2();
+                await createAsyncPromise(10);
+                this.mutate.mut1();
+                this.innerAction();
+                this.mutate.mut2();
+
+                setTimeout(() => {
+                    console.log(eventActionTypes);
+                    assert.deepEqual(eventActionTypes, [
+                        "myStore # @@INIT",
+                        "myStore # action1 / mut1",
+                        "myStore # innerAction / mut1",
+                        "myStore # innerAction / mut2",
+                        "myStore # action1 / mut2",
+                        "myStore # action1 / mut1",
+                        "myStore # innerAction / mut1",
+                        "myStore # innerAction / mut2",
+                        "myStore # action1 / mut2",
+                    ]);
+                    done();
+                }, 0);
+            }
+
+            private innerAction() {
+                this.mutate.mut1();
+                this.mutate.mut2();
+            }
+        }
+
+        globalStateChanges$.subscribe((event) => {
+            eventActionTypes.push(event.storeName + " # " + event.action.type);
+        });
+
+        const store = new MyStore("myStore", new MyMutators(), {});
+        store.action1();
+    });
+
 
 });